+++ flexmeasures/data/schemas/__init__.py
@@ -10,6 +10,8 @@ from .sources import DataSourceIdField as SourceIdField
 from .times import (
     AwareDateTimeField,
     DurationField,
+    FixedDurationField,
+    NominalDurationField,
     ResolutionField,
     TimeIntervalField,
     StartEndTimeSchema,